refactor: In Domain's model.game's openings -> model.game.opening.
