+++ app/src/main/java/no/usn/mob3000/data/repository/game/OpeningsRepository.kt
@@ -3,15 +3,12 @@ package no.usn.mob3000.data.repository.game
 import android.util.Log
 import io.github.jan.supabase.SupabaseClient
 import io.github.jan.supabase.postgrest.postgrest
-import io.github.jan.supabase.postgrest.rpc
-import kotlinx.serialization.json.Json
 import no.usn.mob3000.data.model.game.OpeningsDto
 import no.usn.mob3000.data.network.SupabaseClientWrapper
 import no.usn.mob3000.data.source.remote.game.OpeningsDataSource
 import no.usn.mob3000.domain.helper.game.mapToDomain
-import no.usn.mob3000.domain.model.game.Opening
+import no.usn.mob3000.domain.model.game.opening.Opening
 import no.usn.mob3000.domain.repository.game.IOpeningsRepository
-import java.util.UUID
 
 /**
  * The openings repository orchestrates opening-related data operations between
