+++ app/src/main/java/no/usn/mob3000/domain/model/game/group/GroupStruct.kt
@@ -2,7 +2,7 @@ package no.usn.mob3000.domain.model.game.group
 
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
-import no.usn.mob3000.domain.model.game.Opening
+import no.usn.mob3000.domain.model.game.opening.Opening
 import java.util.UUID
 
 /**
