+++ app/src/main/java/no/usn/mob3000/domain/usecase/game/opening/FetchOpeningsUseCase.kt
@@ -1,6 +1,6 @@
 package no.usn.mob3000.domain.usecase.game.opening
 
-import no.usn.mob3000.domain.model.game.Opening
+import no.usn.mob3000.domain.model.game.opening.Opening
 import no.usn.mob3000.domain.repository.game.IOpeningsRepository
 
 /**
