+++ app/src/main/java/no/usn/mob3000/ui/screens/chess/train/group/CreateGroupScreen.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.flow.StateFlow
 import no.usn.mob3000.R
 import no.usn.mob3000.domain.model.auth.state.AuthenticationState
 import no.usn.mob3000.domain.model.game.group.Group
-import no.usn.mob3000.domain.model.game.Opening
+import no.usn.mob3000.domain.model.game.opening.Opening
 import no.usn.mob3000.ui.components.game.group.GroupEditor
 import no.usn.mob3000.ui.components.settings.SettingsSectionAdmin
