+++ app/src/test/java/no/usn/mob3000/game/OpeningMapperUnitTest.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.json.put
 import no.usn.mob3000.data.model.game.OpeningsDto
 import no.usn.mob3000.domain.helper.game.mapToData
 import no.usn.mob3000.domain.helper.game.mapToDomain
-import no.usn.mob3000.domain.model.game.Opening
+import no.usn.mob3000.domain.model.game.opening.Opening
 import org.junit.Test
 import org.junit.Assert.*
